# backend/docs/swagger.yaml
definitions:
  github_com_MattBrs_OcelotMDM_internal_api_dto_device_dto.DeviceCreationRequest:
    properties:
      architecture:
        type: string
      otp:
        type: string
      type:
        type: string
    required:
    - architecture
    - otp
    - type
    type: object
  github_com_MattBrs_OcelotMDM_internal_api_dto_device_dto.DeviceCreationResponse:
    properties:
      name:
        description: 'TODO: add needed client certificates and configurations in the
          response'
        type: string
      ovpn_file:
        type: string
    required:
    - name
    - ovpn_file
    type: object
  github_com_MattBrs_OcelotMDM_internal_api_dto_device_dto.UpdateAddressRequest:
    properties:
      ip_address:
        type: string
      name:
        type: string
    required:
    - ip_address
    - name
    type: object
  github_com_MattBrs_OcelotMDM_internal_api_dto_token_dto.NewTokenResponse:
    properties:
      expires_at:
        type: string
      token:
        type: string
    type: object
  github_com_MattBrs_OcelotMDM_internal_api_dto_token_dto.NewTokenResponseErr:
    properties:
      error:
        type: string
    type: object
  github_com_MattBrs_OcelotMDM_internal_api_dto_user_dto.CreateUserRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  github_com_MattBrs_OcelotMDM_internal_api_dto_user_dto.CreateUserResponse:
    properties:
      error:
        type: string
      username:
        type: string
    type: object
  github_com_MattBrs_OcelotMDM_internal_api_dto_user_dto.LoginUserRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  github_com_MattBrs_OcelotMDM_internal_api_dto_user_dto.LoginUserResponse:
    properties:
      error:
        type: string
      token:
        type: string
    type: object
  github_com_MattBrs_OcelotMDM_internal_api_dto_user_dto.UpdateUserEnableStatusRequest:
    properties:
      enabled:
        type: boolean
      username:
        type: string
    required:
    - enabled
    - username
    type: object
  github_com_MattBrs_OcelotMDM_internal_api_dto_user_dto.UpdateUserEnableStatusResponseErr:
    properties:
      error:
        type: string
    type: object
  github_com_MattBrs_OcelotMDM_internal_domain_device.Device:
    properties:
      architecture:
        type: string
      id:
        type: string
      ipaddress:
        type: string
      lastSeen:
        type: integer
      name:
        type: string
      status:
        type: string
      supportedCommands:
        items:
          type: string
        type: array
      tags:
        items:
          type: string
        type: array
      type:
        type: string
    type: object
info:
  contact: {}
paths:
  /devices:
    get:
      consumes:
      - application/json
      description: show a filtered list of the devices previously enrolled
      parameters:
      - description: Device ID
        in: query
        name: id
        type: string
      - description: Device Status
        in: query
        name: status
        type: string
      - description: Device Name
        in: query
        name: name
        type: string
      - description: Device Architecture
        in: query
        name: architecture
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MattBrs_OcelotMDM_internal_domain_device.Device'
            type: array
      security:
      - JWT: []
      summary: show a filtered list of the devices
      tags:
      - devices
    post:
      consumes:
      - application/json
      description: enroll device
      parameters:
      - description: Device Data
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/github_com_MattBrs_OcelotMDM_internal_api_dto_device_dto.DeviceCreationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MattBrs_OcelotMDM_internal_api_dto_device_dto.DeviceCreationResponse'
      summary: enroll a new device to the network
      tags:
      - devices
  /devices/updateAddress:
    post:
      consumes:
      - application/json
      description: update known device ip address
      parameters:
      - description: Device New address
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/github_com_MattBrs_OcelotMDM_internal_api_dto_device_dto.UpdateAddressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MattBrs_OcelotMDM_internal_api_dto_device_dto.DeviceCreationResponse'
      security:
      - JWT: []
      summary: update device address
      tags:
      - devices
  /token/generate:
    post:
      consumes:
      - application/json
      description: generates an OTP to be used for some operations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MattBrs_OcelotMDM_internal_api_dto_token_dto.NewTokenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MattBrs_OcelotMDM_internal_api_dto_token_dto.NewTokenResponseErr'
      security:
      - JWT: []
      summary: Generates OTP
      tags:
      - token
  /user/create:
    post:
      consumes:
      - application/json
      description: Create a new user account with username and password. Default disabled
      parameters:
      - description: User account
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_MattBrs_OcelotMDM_internal_api_dto_user_dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MattBrs_OcelotMDM_internal_api_dto_user_dto.CreateUserResponse'
      summary: Create a new User account
      tags:
      - user
  /user/login:
    post:
      consumes:
      - application/json
      description: Lets a user login with his credentials to obtain a JWT token to
        use in other API calls
      parameters:
      - description: User account
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_MattBrs_OcelotMDM_internal_api_dto_user_dto.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MattBrs_OcelotMDM_internal_api_dto_user_dto.LoginUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_MattBrs_OcelotMDM_internal_api_dto_user_dto.LoginUserResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MattBrs_OcelotMDM_internal_api_dto_user_dto.LoginUserResponse'
      summary: Login a User account. Provides JWT
      tags:
      - user
  /user/update/enabled:
    post:
      consumes:
      - application/json
      description: Lets a user login with his credentials to obtain a JWT token to
        use in other API calls
      parameters:
      - description: User account
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_MattBrs_OcelotMDM_internal_api_dto_user_dto.UpdateUserEnableStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MattBrs_OcelotMDM_internal_api_dto_user_dto.UpdateUserEnableStatusRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_MattBrs_OcelotMDM_internal_api_dto_user_dto.UpdateUserEnableStatusResponseErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MattBrs_OcelotMDM_internal_api_dto_user_dto.UpdateUserEnableStatusResponseErr'
      security:
      - JWT: []
      summary: Login a User account. Provides JWT
      tags:
      - user
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"